Move movies list logic to presenter
